requirements.txt


streamlit==1.29.0
audio-recorder-streamlit==0.0.3  # ✅ added for mic input support
streamlit-webrtc==0.45.0         # ✅ pinned for stability
openai>=1.0.0
python-dotenv>=1.0.1
requests>=2.31.0
pypdf>=3.16.4
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
scipy>=1.10.0
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
azure-cognitiveservices-speech>=1.34.1
gtts>=2.3.2
webrtcvad>=2.0.10
beautifulsoup4



